+++ src/features/post/hooks/useFilterCategory.ts
@@ -1,0 +1,47 @@
+import { useCallback, useEffect, useState } from 'react';
+
+import { getPostsByCategory } from '@/features/post';
+import { useFilterStore } from '@/features/post/store/useFilterStore';
+
+type CategoryItem = { id: string; label: string };
+type SettingsMap = Record<string, boolean>;
+
+export function useFilterCategory() {
+  const [categories, setCategories] = useState<CategoryItem[]>([]);
+  const [checked, setChecked] = useState<string[]>([]);
+
+  const { filter, setFilter } = useFilterStore();
+
+  const fetchCategories = useCallback(async () => {
+    const postsByCategory = await getPostsByCategory();
+    const keys = Object.keys(postsByCategory);
+
+    setCategories(keys.map((key) => ({ id: key, label: key })));
+
+    if (!filter) {
+      const initialFilter = Object.fromEntries(keys.map((key) => [key, true])) as SettingsMap;
+      await setFilter(initialFilter);
+      setChecked(keys);
+    } else {
+      setChecked(keys.filter((key) => filter[key]));
+    }
+  }, [filter, setFilter]);
+
+  useEffect(() => {
+    fetchCategories();
+  }, [fetchCategories]);
+
+  const saveSettings = async (nextChecked: string[]) => {
+    const next: SettingsMap = Object.fromEntries(
+      categories.map((c) => [c.id, nextChecked.includes(c.id)]),
+    );
+    await setFilter(next);
+    setChecked(nextChecked);
+  };
+
+  return {
+    categories,
+    checked,
+    saveSettings,
+  };
+}
