+++ src/features/post/index.ts
@@ -2,6 +2,8 @@ export * from './components/MDXHeader';
 export * from './components/MDXRender';
 export * from './components/PostSidebarLayout';
 export * from './components/RenderPost';
+export * from './hooks/useFilterCategory';
 export * from './hooks/usePostList';
+export * from './store/useFilterStore';
 export * from './types/postTypes';
 export * from './utils/postUtils';
