+++ src/hooks/index.ts
@@ -1,4 +1,3 @@
-export * from './useFilterCategory';
 export * from './useMobile';
 export * from './useTheme';
 export * from './useToast';
